fix: declare SSM/DeltaNet parameters with build dtype to prevent SIGABRT
When a model is built with dtype=FLOAT16, HuggingFace stores all weights as
float16. Components that intentionally upcast parameters to float32 (for
numerical stability) used nn.Parameter() with the default FLOAT type
annotation. The constant folding pass then saw Cast(param, to=FLOAT) where
input_type=FLOAT and output_type=FLOAT and eliminated the Cast as an identity.
This left float16 const_value data labeled with a float32 type annotation in
the ONNX proto, causing ORT's CUDA EP to crash with a vector bounds assertion
(stl_vector.h:1130: '__n < this->size()').

Fix: declare dt_bias, A_log, D, and dt_bias parameters with dtype=config.dtype
so their type annotation matches the actual HF weight dtype. With the correct
annotation, Cast(FLOAT16, to=FLOAT) is seen as a type change and preserved by
constant folding, correctly producing float32 values.

Affected components:
- GatedDeltaNet: dt_bias, A_log (SIGABRT root cause for Qwen3.5)
- SelectiveScan: A_log, D (proactive fix for Mamba1/Jamba)
- JambaSelectiveScan: inherits dtype via SelectiveScan
- Mamba2Scan: A_log, D, dt_bias (proactive fix for Mamba2/Bamba/NemotronH)
- GatedRMSNorm: weight (proactive fix for CUDA workaround path in NemotronH)

+++ src/mobius/components/_gated_deltanet.py
@@ -138,9 +138,13 @@ class GatedDeltaNet(nn.Module):
         # Causal depthwise Conv1D
         self.conv1d = _DepthwiseConv1d(self.conv_dim, self.conv_kernel_size)
 
-        # Learnable parameters for decay computation
-        self.dt_bias = nn.Parameter([self.num_v_heads])
-        self.A_log = nn.Parameter([self.num_v_heads])
+        # Learnable parameters for decay computation.
+        # Use config.dtype so the type annotation matches the HF weight dtype.
+        # Without this, Cast(dt_bias, to=FLOAT) has input_type=output_type=FLOAT
+        # (from the default annotation) and is eliminated as identity by constant
+        # folding, leaving float16 data with a float32 type label → ORT SIGABRT.
+        self.dt_bias = nn.Parameter([self.num_v_heads], dtype=config.dtype)
+        self.A_log = nn.Parameter([self.num_v_heads], dtype=config.dtype)
 
         # Gated output normalization
         self.norm = PostGatedRMSNorm(self.head_v_dim, eps=config.rms_norm_eps)
